+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## Version [2.1.2] - 2025-04-20
+### New features (Nuevo)
+### Improvements (Mejoras)
+- Caché para imágenes con Service Worker y Proxy
+- Imágen de marca en las tarjetas de items
+- Optimización de carga de imágenes, lazy
+- Mejora visual en la tarjeta del producto de la grilla, elevación de la misma con ratón y bordes
+- Eliminación de Zoom en las imágenes
+- Proxy 
+### Bug fixes (Corección de Errores)
+- Arreglo de recuperación de lista de precios de la sesión, valida si es nulo o no
+### Tasks
+
 ## Version [2.1.1] - 2024-08-30
 ### New features (Nuevo)
 ### Improvements (Mejoras)